refactor(sequencer/feeds): Change Feed ID size to 128b

+++ apps/oracles/eth-rpc/src/lib.rs
@@ -57,7 +57,7 @@ pub struct ResponseEthCallError {
 #[derive(Deserialize, Debug, Clone)]
 pub struct ResponseEthCall {
     pub jsonrpc: String,
-    pub id: Option<u64>,
+    pub id: Option<u128>,
     pub error: Option<ResponseEthCallError>,
     pub result: Option<String>,
     #[serde(default)]
@@ -126,7 +126,7 @@ impl RequestEthCall {
     }
 }
 
-type FetchedDataForFeed = HashMap<u64, ResponseEthCall>;
+type FetchedDataForFeed = HashMap<u128, ResponseEthCall>;
 type MyProvider = alloy::providers::fillers::FillProvider<
     alloy::providers::fillers::JoinFill<
         alloy::providers::Identity,
@@ -222,7 +222,7 @@ fn process_results(results: &FetchedDataForFeed) -> Result<Payload> {
 }
 
 fn print_results(results: &FetchedDataForFeed, payload: &Payload) {
-    let mut keys = results.keys().cloned().collect::<Vec<u64>>();
+    let mut keys = results.keys().cloned().collect::<Vec<u128>>();
     keys.sort();
 
     let mut table = Table::new();
@@ -278,7 +278,7 @@ pub struct Pair {
 #[derive(Deserialize, Debug)]
 struct FeedConfig {
     #[serde(default)]
-    pub feed_id: u64,
+    pub feed_id: u128,
     pub arguments: Vec<String>,
 }
 
@@ -287,7 +287,7 @@ fn get_resources_from_settings(settings: &Settings) -> Result<Vec<FeedConfig>> {
     for feed_setting in &settings.data_feeds {
         let mut feed_config = serde_json::from_str::<FeedConfig>(&feed_setting.data)
             .context("Couldn't parse data feed")?;
-        feed_config.feed_id = feed_setting.id.parse::<u64>()?;
+        feed_config.feed_id = feed_setting.id.parse::<u128>()?;
         config.push(feed_config);
     }
     Ok(config)
